+++ src/extensions/init/initext_hooks.cpp
@@ -51,8 +51,10 @@
 #include <Windows.h>
 #include <commctrl.h>
 
+#include "extension.h"
 #include "hooker.h"
 #include "hooker_macros.h"
+#include "sideext.h"
 
 
 extern HMODULE DLLInstance;
@@ -633,6 +635,14 @@ bool Vinifera_Prep_For_Side(SideType side)
     }
 
     Map.Init_For_House();
+
+    /**
+     *  Set the options menu color.
+     */
+    static auto& OptionsColor = Make_Global<COLORREF>(0x00808B7C);
+    const SideClassExtension* sideext = Extension::Fetch<SideClassExtension>(Sides[static_cast<SideType>(Scen->IsGDI)]);
+    const RGBStruct& options_rgb = sideext->OptionsColor;
+    OptionsColor = RGB(options_rgb.R, options_rgb.G, options_rgb.B);
 
     return true;
 }